final cannon
still an issue with targeting, the big fat mushroom head is in the way.
ultimately, this needs a point and click to target, showing the mouse cursor, and highlighting targeted object. as it stands we have something halfway between 3rd and 1st person.
or perhaps a gun was the wrong weapon choice. design needs to be informed by implementation as much as implementation is informed by design.